* Bump version

* fix(935): Update default size for s2-l1-shared volume

* fix(924): Update default limits.memory

* chore: update lib and add catalog events

* fix(928): Use metadata field datastripId instead of productGroupId in order to update L0GR or L1CTL availability

* fix(923): Add timeout for obs transfers

* fix(923): Add global retry for OBS transfers + doc

* ci: update doc export

* ci: fix again

* docs: fix typos in Readme

* fix(901): Remove RESTART allowed Action for L2PW (#33)

* fix(945): fixed upload bug with 0 file (#34)

* fix(917) : Double L0c DS production generates fake alarms (#35)

Increase kafka timeout from 1 hour to 2 hours.

* fix(941): Updated MissingOutput behaviour to be correct

* fix(941): Added configurable variables for IPF versions and default number of missing outputs

* feat: Update IPF version

* chore: code cleanup

* fix(970): add cleanup for aux files before download

* docs: Update conf and docs with latest OPS values

* Prepare for release 1.7.0-rc1

* Bump version

* fix: corrected Processing Station default value for L2 orchestrator (#40)

* Chore/memory and cpu update (#41)

chore: memory and cpu values update to match rs-config values

* fix(999): added custom field to PW Tracelogs (#42)

* doc: added limits and requests for filters and routers (#43)

* Update README.md

Homogenization of the titles of the documentation

* Update README.md

Homogenization of the titles of the documentation

* Update README.md

Homogenization of the titles of the documentation

* ci: push and remove images after security scan, in order to save space (#45)

* ci: fix typos

* doc: Add resubmit procedure (#46)

* Update docker-build-release.yml

Added REGISTRY_PROJECT in the image name for the part "Push and clean docker image"

* docs: Update NOTICE.md

* docs(897): Explain how to freeze Processing Baseline (#48)

* ci: Update docker-build-release.yml

* Prepare for release 1.8.0-rc1

---------

+++ apps/libs/core-obs/src/main/java/eu/csgroup/coprs/ps2/core/obs/service/ObsService.java
@@ -162,7 +162,8 @@ public class ObsService {
 
         log.debug("Downloading {} folders using FileInfos", fileInfoSet.size());
 
-        Mono.when(fileInfoSet.stream()
+        waitOnTransfers(
+                fileInfoSet.stream()
                         .map(fileInfo -> {
                             final String key = fileInfo.getKey();
                             final String bucket = fileInfo.getBucket();
@@ -173,8 +174,9 @@ public class ObsService {
                                 return doDirDownload(key, bucket, destinationPath);
                             }
                         })
-                        .toList())
-                .block();
+                        .toList(),
+                obsProperties.getDownloadTimeout()
+        );
     }
 
     public void upload(Set<FileInfo> fileInfoSet, UUID parentUid) {
@@ -190,7 +192,8 @@ public class ObsService {
 
         log.debug("Uploading {} files using FileInfos", fileInfoSet.size());
 
-        Mono.when(fileInfoSet.stream()
+        waitOnTransfers(
+                fileInfoSet.stream()
                         .map(fileInfo -> {
                             final Path sourcePath = Paths.get(fileInfo.getFullLocalPath());
                             if (sourcePath.toFile().isDirectory()) {
@@ -199,8 +202,8 @@ public class ObsService {
                                 return doFileUpload(sourcePath, fileInfo.getBucket(), fileInfo.getKey());
                             }
                         })
-                        .toList())
-                .block();
+                        .toList(),
+                obsProperties.getUploadTimeout());
     }
 
     /**
@@ -212,52 +215,68 @@ public class ObsService {
      */
     public void uploadWithMd5(Set<FileInfo> fileInfoSet, UUID parentUid) {
 
-        log.info("Uploading {} folders to OBS", fileInfoSet.size());
-        upload(fileInfoSet, parentUid);
+        if (!fileInfoSet.isEmpty()) {
 
-        log.info("Creating md5sum files");
+            log.info("Uploading {} folders to OBS", fileInfoSet.size());
+            upload(fileInfoSet, parentUid);
 
-        String tmpFolder = FolderParameters.TMP_DOWNLOAD_FOLDER + FileSystems.getDefault().getSeparator() + UUID.randomUUID();
-        FileOperationUtils.createFolders(Set.of(tmpFolder));
+            log.info("Creating md5sum files");
 
-        Set<FileInfo> md5FileInfos = new HashSet<>();
+            String tmpFolder = FolderParameters.TMP_DOWNLOAD_FOLDER + FileSystems.getDefault().getSeparator() + UUID.randomUUID();
+            FileOperationUtils.createFolders(Set.of(tmpFolder));
 
-        fileInfoSet.forEach(fileInfo -> {
+            Set<FileInfo> md5FileInfos = new HashSet<>();
 
-            final Path localPath = Paths.get(fileInfo.getFullLocalPath());
-            final Path md5sumPath = Paths.get(tmpFolder).resolve(localPath.getFileName() + MD5SUM_SUFFIX);
+            fileInfoSet.forEach(fileInfo -> {
 
-            final Map<String, String> md5ByFileName = Md5Utils.getMd5(localPath);
-            final Map<String, String> eTagByKey = getETags(fileInfo.getBucket(), fileInfo.getKey());
+                final Path localPath = Paths.get(fileInfo.getFullLocalPath());
+                final Path md5sumPath = Paths.get(tmpFolder).resolve(localPath.getFileName() + MD5SUM_SUFFIX);
 
-            final List<String> lines = md5ByFileName.entrySet()
-                    .stream()
-                    .map(entry -> String.format("%s %s %s", entry.getValue(), eTagByKey.get(entry.getKey()), entry.getKey()))
-                    .toList();
+                final Map<String, String> md5ByFileName = Md5Utils.getMd5(localPath);
+                final Map<String, String> eTagByKey = getETags(fileInfo.getBucket(), fileInfo.getKey());
 
-            try (final OutputStream outputStream = Files.newOutputStream(md5sumPath)) {
-                for (String line : lines) {
-                    outputStream.write(line.getBytes());
-                    outputStream.write(System.lineSeparator().getBytes());
+                final List<String> lines = md5ByFileName.entrySet()
+                        .stream()
+                        .map(entry -> String.format("%s %s %s", entry.getValue(), eTagByKey.get(entry.getKey()), entry.getKey()))
+                        .toList();
+
+                try (final OutputStream outputStream = Files.newOutputStream(md5sumPath)) {
+                    for (String line : lines) {
+                        outputStream.write(line.getBytes());
+                        outputStream.write(System.lineSeparator().getBytes());
+                    }
+                } catch (IOException e) {
+                    throw new FileOperationException("Unable to create temporary files in: " + tmpFolder, e);
                 }
-            } catch (IOException e) {
-                throw new FileOperationException("Unable to create temporary files in: " + tmpFolder, e);
+
+                md5FileInfos.add(
+                        new FileInfo()
+                                .setFullLocalPath(md5sumPath.toString())
+                                .setBucket(fileInfo.getBucket())
+                                .setKey(md5sumPath.getFileName().toString())
+                );
+            });
+
+            log.info("Uploading md5sum files");
+            upload(md5FileInfos, parentUid);
+
+            if (!FileUtils.deleteQuietly(new File(tmpFolder))) {
+                log.warn("Unable to delete temp folder {}", tmpFolder);
             }
-
-            md5FileInfos.add(
-                    new FileInfo()
-                            .setFullLocalPath(md5sumPath.toString())
-                            .setBucket(fileInfo.getBucket())
-                            .setKey(md5sumPath.getFileName().toString())
-            );
-        });
-
-        log.info("Uploading md5sum files");
-        upload(md5FileInfos, parentUid);
-
-        if (!FileUtils.deleteQuietly(new File(tmpFolder))) {
-            log.warn("Unable to delete temp folder {}", tmpFolder);
         }
+    }
+
+    private void waitOnTransfers(List<Mono<?>> transfers, int timeout) {
+        Mono.when(transfers)
+                .timeout(
+                        Duration.ofMinutes(timeout),
+                        Mono.fromCallable(() -> {
+                            final String message = String.format("Transfer operation timed out after %s minutes", timeout);
+                            log.error(message);
+                            throw new ObsException(message);
+                        }))
+                .retryWhen(getRetrySpec())
+                .block();
     }
 
     private Mono<?> doFileDownload(String key, String bucket, Path destinationPath) {
